Refactor: clean up and refactor auth api
