+++ src/auth/constants.ts
@@ -1,0 +1,4 @@
+export const TOKEN_STORAGE = 'colony-server-token';
+
+export const URL_AUTH_TOKEN = `${process.env.SERVER_ENDPOINT}/auth/token`;
+export const URL_AUTH_CHALLENGE = `${process.env.SERVER_ENDPOINT}/auth/challenge`;
